Add the device auth flow via .device_connect() method
This is an alternative flow that requires the user to follow a link
and confirm a code but allows the app to run in a separate context
such as from the command line.

Also added ability to optionally print a QR code if the "qrcode"
module is installed, to make it easier to authenticate with a mobile
device.